Proper handling of multi-line responses.
The first line begins with exact required reply code followed by a
hypen. A client should then simply search for the second occurence of
the same reply code immediately followed by space at the beginning of
the line. The server must pad the front of a line beginning with a
3-digit number to avoid confusion.