Unlink and gc custom io.
After

1. reading the caveats of using stdio with socket io from #14.8
from "The Unix Networking programming"

2. develop common io for (now foregone) FTP interactive interpreter
and rest of the code (clearing of buffered data on EINTR for
interrupted commands)

a custom io layer was developed.

In the hingsight, stdio is much easier (the only caveat is to flush
between a read and a write). Also discussed with eric@ when a similar
question of using "stdio with sockets" arose while reviewing
delivery_lmtp.c from smtpd(8).

jca@ suggested using stdio when I showed it during g2k16 and gilles@
as well when he reviewed an earlier version of this code. deraadt@
disliked io.c as well.